requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.3.0
DateTime==5.5
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.30.1
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.4.2
lxml==5.3.1
lxml_html_clean==0.4.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
newsapi-python==0.2.7
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
safetensors==0.5.3
setuptools==78.1.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.6
sympy==1.13.1
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.3
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
zope.interface==7.2
